requirements.txt
Django>=3.2,<4.0
django-allauth>=0.54.0
django-cors-headers>=3.7.0
djangorestframework>=3.12,<4.0
drf-yasg>=1.21.7
python-decouple>=3.5
requests>=2.25.1
solana>=0.24.0
python-dotenv==1.0.0
PyJWT==2.8.0
cryptography==42.0.4  # 修改版本号

# 开发依赖
black==25.1.0
flake8==7.1.2
pylint==3.3.6
pytest==8.0.0
pytest-django==4.8.0
